+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+
+* [PR-18](https://github.com/AGH-CEAI/aegis_ros/pull/18) - Added scripts for running the project in `production` mode or in `development` toolbx container.
+
 ### Changed
 ### Deprecated
 ### Removed